refactor: ContributorTopAppBar 함수를 별도의 파일로 분리하고 top 여부를 전달받도록 변경
